FIX Generator indexing
Generator indexing returned empty generator because generator is used
twice in the slicerator code (having to do with the inrange checking).
Moved the inrange checking.

Test passed because in `assert_letters_equal`, test passed if one of the
values where empty. Now lengths are checked explicitly (so if testing a
generator, it has to be converted to a list).